refactor: simplify RAG pipeline and update docs
- Remove legacy build_knowledge.py and data/processed/ directory
- Delete orphaned geos_docs collection from ChromaDB
- Update README with current dual-collection pipeline:
  - parse_rst_chunks.py → build_vector_index.py workflow
  - RAG architecture documentation (Navigator + Technical collections)
  - Shadow embeddings and lazy code loading explanation
- Add peek_vector_db.py utility script for debugging
- Fix PROJECT_ROOT path in scripts/tests/
